cleanup - removed section page wrappers, layout tweaks - sub nav fixed width, main content auto width and centered
