+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "adcp-ui",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "description": "UI for ADCP.",
   "scripts": {
     "lint": "eslint --ext .js .",
@@ -51,6 +51,7 @@
     "@testing-library/user-event": "^13.5.0",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
+    "react-dropzone": "^12.0.5",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.0",
     "uniqid": "^5.4.0"
